+++ Tests/Core/Registry/RegistryManager.cs
@@ -5,7 +5,6 @@ using NUnit.Framework;
 
 using CKAN;
 using CKAN.Versioning;
-using CKAN.Games.KerbalSpaceProgram;
 
 using Tests.Data;