+++ Tests/GUI/GUIConfiguration.cs
@@ -1,11 +1,10 @@
 using System.IO;
-using System.Collections.Generic;
 
 using NUnit.Framework;
 
 using CKAN;
+using CKAN.IO;
 using CKAN.GUI;
-
 using Tests.Data;
 
 namespace Tests.GUI
@@ -13,49 +12,48 @@ namespace Tests.GUI
     [TestFixture]
     public class GuiConfigurationTests
     {
-        private string? tempDir;
-
-        [OneTimeSetUp]
-        public void Setup()
-        {
-            tempDir = TestData.NewTempDir();
-        }
-
-        [OneTimeTearDown]
-        public void TearDown()
-        {
-            if (tempDir != null)
-            {
-                Directory.Delete(tempDir, true);
-            }
-        }
-
         [Test]
         public void LoadOrCreateConfiguration_MalformedXMLFile_ThrowsKraken()
         {
-            string tempFile = Path.Combine(tempDir!, "invalid.xml");
-
-            using (var stream = new StreamWriter(tempFile))
+            // Arrange
+            using (var inst = new DisposableKSP())
             {
-                stream.Write("This is not a valid XML file.");
-            }
+                var xmlPath = Path.Combine(inst.KSP.CkanDir(), "GUIConfig.xml");
+                using (var stream = new StreamWriter(Path.Combine(inst.KSP.CkanDir(),
+                                                                  "GUIConfig.xml")))
+                {
+                    stream.Write("This is not a valid XML file.");
+                }
+                var steamLib = new SteamLibrary(null);
 
-            Assert.Throws<Kraken>(() => GUIConfiguration.LoadOrCreateConfiguration(tempFile, new List<string>()));
+                // Act / Assert
+                Assert.Throws<Kraken>(() => GUIConfiguration.LoadOrCreateConfiguration(inst.KSP, steamLib));
+                Assert.IsTrue(File.Exists(xmlPath));
+            }
         }
 
         [Test]
         public void LoadOrCreateConfiguration_CorrectConfigurationFile_Loaded()
         {
-            string tempFile = Path.Combine(tempDir!, "valid.xml");
-
-            using (var stream = new StreamWriter(tempFile))
+            // Arrange
+            using (var inst = new DisposableKSP())
             {
-                stream.Write(TestData.ConfigurationFile());
+                var xmlPath  = Path.Combine(inst.KSP.CkanDir(), "GUIConfig.xml");
+                var jsonPath = Path.Combine(inst.KSP.CkanDir(), "GUIConfig.json");
+                using (var stream = new StreamWriter(xmlPath))
+                {
+                    stream.Write(TestData.ConfigurationFile());
+                }
+                var steamLib = new SteamLibrary(null);
+
+                // Act / Assert
+                var result = GUIConfiguration.LoadOrCreateConfiguration(inst.KSP, steamLib);
+                Assert.IsNotNull(result);
+                Assert.AreEqual(512, result.WindowLoc.X);
+                Assert.AreEqual(136, result.WindowLoc.Y);
+                Assert.IsFalse(File.Exists(xmlPath));
+                Assert.IsTrue(File.Exists(jsonPath));
             }
-
-            var result = GUIConfiguration.LoadOrCreateConfiguration(tempFile, new List<string>());
-
-            Assert.IsNotNull(result);
         }
     }
 }